Fix ignored & move to obscure
