+++ plugin/src/main/resources/META-INF/plugin.xml
@@ -39,22 +39,6 @@
                          groupName="PyNose"
                          enabledByDefault="true"
                          level="WARNING"
-                         implementationClass="org.jetbrains.research.pynose.plugin.inspections.unittest.IgnoredTestTestSmellUnittestInspection"
-                         displayName="Ignored test unittest"
-                         shortName="IgnoredTestTestSmellUnittestInspection"/>
-        <localInspection language="Python"
-                         groupPath="Python"
-                         groupName="PyNose"
-                         enabledByDefault="true"
-                         level="WARNING"
-                         implementationClass="org.jetbrains.research.pynose.plugin.inspections.pytest.IgnoredTestTestSmellPytestInspection"
-                         displayName="Ignored test pytest"
-                         shortName="IgnoredTestTestSmellPytestInspection"/>
-        <localInspection language="Python"
-                         groupPath="Python"
-                         groupName="PyNose"
-                         enabledByDefault="true"
-                         level="WARNING"
                          implementationClass="org.jetbrains.research.pynose.plugin.inspections.disabled.ConditionalTestLogicTestSmellInspection"
                          displayName="Conditional logic test"
                          shortName="ConditionalTestLogicTestSmellInspection"/>
