+++ plugin/src/test/kotlin/org/jetbrains/research/pynose/plugin/inspections/obscure/IgnoredTestTestSmellPytestInspectionTests.kt
@@ -1,37 +1,39 @@
-package org.jetbrains.research.pynose.plugin.inspections.unittest
+package org.jetbrains.research.pynose.plugin.inspections.obscure
 
 import com.intellij.lang.annotation.HighlightSeverity
 import com.intellij.openapi.components.service
 import io.mockk.every
 import io.mockk.mockkObject
 import org.jetbrains.research.pynose.plugin.inspections.TestRunnerServiceFacade
+import org.jetbrains.research.pynose.plugin.startup.PyNoseMode
 import org.jetbrains.research.pynose.plugin.util.AbstractTestSmellInspectionTestWithSdk
 import org.jetbrains.research.pynose.plugin.util.TestSmellBundle
 import org.junit.Test
 import org.junit.jupiter.api.BeforeAll
 
-class IgnoredTestTestSmellUnittestInspectionTests : AbstractTestSmellInspectionTestWithSdk() {
+class IgnoredTestTestSmellPytestInspectionTests : AbstractTestSmellInspectionTestWithSdk() {
 
     @BeforeAll
     override fun setUp() {
         super.setUp()
+        mockkObject(PyNoseMode)
         mockkObject(myFixture.project.service<TestRunnerServiceFacade>())
         every {
             myFixture.project.service<TestRunnerServiceFacade>().getConfiguredTestRunner(any())
-        } returns "Unittests"
-        myFixture.enableInspections(IgnoredTestTestSmellUnittestInspection())
+        } returns "pytest"
+        myFixture.enableInspections(IgnoredTestTestSmellPytestInspection())
     }
 
     override fun getTestDataPath(): String {
-        return "src/test/resources/org/jetbrains/research/pynose/plugin/inspections/data/ignored/unittest"
+        return "src/test/resources/org/jetbrains/research/pynose/plugin/inspections/data/ignored/pytest"
     }
 
     @Test
-    fun `test skipped without unittest dependency`() {
+    fun `test skipped wrong class name`() {
         myFixture.configureByText(
-            "test_file.py", "import unittest\n" +
-                    "class SomeClass:\n" +
-                    "    @unittest.skip(\"reason\")\n" +
+            "test_file.py", "import pytest\n" +
+                    "class TestClass:\n" +
+                    "    @pytest.mark.skip(reason=\"reason\")\n" +
                     "    def test_something(self):\n" +
                     "        pass"
         )
@@ -40,10 +42,9 @@ class IgnoredTestTestSmellUnittestInspectionTests : AbstractTestSmellInspectionT
     }
 
     @Test
-    fun `test not skipped with unittest dependency`() {
+    fun `test not skipped`() {
         myFixture.configureByText(
-            "test_file.py", "import unittest\n" +
-                    "class SomeTestCase(unittest.TestCase):\n" +
+            "test_file.py", "class SomeTestCase:\n" +
                     "    def test_something(self):\n" +
                     "        pass"
         )
@@ -54,23 +55,22 @@ class IgnoredTestTestSmellUnittestInspectionTests : AbstractTestSmellInspectionT
     @Test
     fun `test skip class with comment`() {
         myFixture.configureByText(
-            "test_file.py", "import unittest\n" +
-                    "@unittest.skip(\"reason\")\n" +
-                    "class SomeClass(unittest.TestCase):\n" +
+            "test_file.py", "import pytest\n" +
+                    "@pytest.mark.skip(reason=\"reason\")\n" +
+                    "class TestClass(unittest.TestCase):\n" +
                     "    def test_something(self):\n" +
                     "        pass"
         )
-        val highlightInfos = myFixture.doHighlighting()
-        assertTrue(!highlightInfos.any { it.severity == HighlightSeverity.WARNING })
+        myFixture.checkHighlighting()
     }
 
     @Test
-    fun `test highlighted skip class`() {
+    fun `test highlighted skipped class`() {
         myFixture.configureByText(
-            "test_file.py", "import unittest\n" +
-                    "@unittest.skip()\n" +
+            "test_file.py", "import pytest\n" +
+                    "@pytest.mark.skip()\n" +
                     "class <warning descr=\"${TestSmellBundle.message("inspections.ignored.description")}\">" +
-                    "SomeClass</warning>(unittest.TestCase):\n" +
+                    "TestClass</warning>():\n" +
                     "    def test_something(self):\n" +
                     "        pass"
         )
@@ -80,22 +80,21 @@ class IgnoredTestTestSmellUnittestInspectionTests : AbstractTestSmellInspectionT
     @Test
     fun `test basic skip with comment`() {
         myFixture.configureByText(
-            "test_file.py", "import unittest\n" +
-                    "class SomeClass(unittest.TestCase):\n" +
-                    "    @unittest.skip(\"reason\")\n" +
+            "test_file.py", "import pytest\n" +
+                    "class TestClass:\n" +
+                    "    @pytest.mark.skip(reason=\"reason\")\n" +
                     "    def test_something(self):\n" +
                     "        pass"
         )
-        val highlightInfos = myFixture.doHighlighting()
-        assertTrue(!highlightInfos.any { it.severity == HighlightSeverity.WARNING })
+        myFixture.checkHighlighting()
     }
 
     @Test
     fun `test highlighted basic skip`() {
         myFixture.configureByText(
-            "test_file.py", "import unittest\n" +
-                    "class SomeClass(unittest.TestCase):\n" +
-                    "    @unittest.skip()\n" +
+            "test_file.py", "import pytest\n" +
+                    "class TestClass:\n" +
+                    "    @pytest.mark.skip()\n" +
                     "    def <warning descr=\"${TestSmellBundle.message("inspections.ignored.description")}\">" +
                     "test_something</warning>(self):\n" +
                     "        pass"
@@ -106,22 +105,21 @@ class IgnoredTestTestSmellUnittestInspectionTests : AbstractTestSmellInspectionT
     @Test
     fun `test skip if with comment`() {
         myFixture.configureByText(
-            "test_file.py", "import unittest\n" +
-                    "class SomeClass(unittest.TestCase):\n" +
-                    "    @unittest.skipIf(mylib.__version__ < (1, 3), \"reason\")\n" +
+            "test_file.py", "import pytest\n" +
+                    "class TestClass:\n" +
+                    "    @pytest.mark.skipIf(mylib.__version__ < (1, 3), reason=\"reason\")\n" +
                     "    def test_something(self):\n" +
                     "        pass"
         )
-        val highlightInfos = myFixture.doHighlighting()
-        assertTrue(!highlightInfos.any { it.severity == HighlightSeverity.WARNING })
+        myFixture.checkHighlighting()
     }
 
     @Test
     fun `test highlighted skip if`() {
         myFixture.configureByText(
-            "test_file.py", "import unittest\n" +
-                    "class SomeClass(unittest.TestCase):\n" +
-                    "    @unittest.skipIf(mylib.__version__ < (1, 3))\n" +
+            "test_file.py", "import pytest\n" +
+                    "class TestClass:\n" +
+                    "    @pytest.mark.skipIf(mylib.__version__ < (1, 3))\n" +
                     "    def <warning descr=\"${TestSmellBundle.message("inspections.ignored.description")}\">" +
                     "test_something</warning>(self):\n" +
                     "        pass"
@@ -130,16 +128,30 @@ class IgnoredTestTestSmellUnittestInspectionTests : AbstractTestSmellInspectionT
     }
 
     @Test
-    fun `test highlighted skip unless`() {
+    fun `test highlighted xfail`() {
         myFixture.configureByText(
-            "test_file.py", "import unittest\n" +
-                    "class SomeClass(unittest.TestCase):\n" +
-                    "    @unittest.skipUnless(mylib.__version__ < (1, 3))\n" +
+            "test_file.py", "import pytest\n" +
+                    "class TestClass:\n" +
+                    "    @pytest.mark.xfail(mylib.__version__ < (1, 3))\n" +
                     "    def <warning descr=\"${TestSmellBundle.message("inspections.ignored.description")}\">" +
                     "test_something</warning>(self):\n" +
                     "        pass"
         )
         myFixture.checkHighlighting()
+    }
+
+    @Test
+    fun `test xfail with comment`() {
+        myFixture.configureByText(
+            "test_file.py", "import pytest\n" +
+                    "class TestClass:\n" +
+                    "    @pytest.mark.xfail(mylib.__version__ < (1, 3), reason=\"reason\")\n" +
+                    "    def <warning descr=\"${TestSmellBundle.message("inspections.ignored.description")}\">" +
+                    "test_something</warning>(self):\n" +
+                    "        pass"
+        )
+        val highlightInfos = myFixture.doHighlighting()
+        assertTrue(!highlightInfos.any { it.severity == HighlightSeverity.WARNING })
     }
 
     @Test
